yatri_backend/requirements.txt
# YatriConnect Backend - Requirements
# Django + DRF + JWT + PostgreSQL + Redis

# Core Django
Django==4.2.27
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# CORS
django-cors-headers==4.3.1

# Security & Validation
python-decouple==3.8
cryptography==41.0.7

# External API Integration
requests==2.31.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
ipython==8.12.3
